# dashboard/swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "A space station except way more Aero-dynamic",
    "title": "Aerostation API service",
    "contact": {
      "email": "sbhagat@aerospike.com"
    },
    "version": "1.0.0"
  },
  "basePath": "/v1",
  "paths": {
...
    "/api/v1/aerospike/clusters": {
      "get": {
        "description": "Get names of all the workload clusters in array of string",
        "produces": [
          "application/json"
        ],
        "tags": [
          "aerospike"
        ],
        "summary": "Get all aerospike database clusters",
        "operationId": "getAerospikeClusters",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "example": [
                "aerospike-cluster-1",
                "aerospike-cluster-2"
              ]
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Create a aerospike database cluster",
        "produces": [
          "application/json"
        ],
        "tags": [
          "aerospike"
        ],
        "summary": "Create a aerospike database cluster",
        "operationId": "createAerospikeCluster",
        "parameters": [
          {
            "description": "Request body for creating workload cluster",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAerospikeClusterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aerospike database created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Unable to create cluster"
          }
        }
      }
    },
    "/api/v1/aerospike/clusters/{name}": {
      "get": {
        "description": "Returns cluster status by name",
        "produces": [
          "application/json"
        ],
        "tags": [
          "aerospike"
        ],
        "summary": "Get database status by name",
        "operationId": "getAerospikeCluster",
        "parameters": [
          {
            "type": "string",
            "description": "Name of the aerospike cluster",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/AeroDatabaseSpec"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "cluster not found"
          }
        }
      },
      "delete": {
        "description": "Delete a aerospike database cluster",
        "produces": [
          "application/json"
        ],
        "tags": [
          "aerospike"
        ],
        "summary": "Delete a aerospike database cluster",
        "operationId": "deleteAerospikeClusters",
        "parameters": [
          {
            "type": "string",
            "description": "Name of the aerospike cluster to delete",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Aerospike cluster deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "description": "Update a aerospike cluster",
        "produces": [
          "application/json"
        ],
        "tags": [
          "aerospike"
        ],
        "summary": "Update a aerospike cluster",
        "operationId": "updateAerospikeClusters",
        "parameters": [
          {
            "type": "string",
            "description": "Name of the aerospike cluster to update",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "description": "Request body for updating aerospike cluster",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DatabaseOptions"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cluster updated successfully"
          },
          "400": {
            "description": "Unable to update aerospike cluster"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/cloud/aws/regins": {
      "get": {
        "description": "Get all of the supported regions in AWS",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cloud"
        ],
        "summary": "Get all supported regions in AWS",
        "operationId": "GetAWSRegions",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/AwsRegionsResponse",
              "example": {
                "regions": [
                  {
                    "continent": "North America",
                    "name": "N. Virginia",
                    "region": "us-east-1"
                  },
                  {
                    "continent": "North America",
                    "name": "N. California",
                    "region": "us-west-1"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/cloud/aws/{region}": {
      "get": {
        "description": "Get all of the supported regions in AWS",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cloud"
        ],
        "summary": "Get all supported regions in AWS",
        "operationId": "GetRegionInstanceTypes",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/AwsRegionsResponse",
              "example": [
                {
                  "regions": {
                    "continent": "North America",
                    "name": "cluster-namespace",
                    "region": "us-east-1"
                  }
                }
              ]
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AeroDatabaseSpec": {
      "description": "AeroDatabaseSpec defines the desired state of AeroDatabase",
      "type": "object",
      "required": [
        "clusterKey",
        "name",
        "namespace",
        "databaseType",
        "options"
      ],
      "properties": {
        "clusterKey": {
          "$ref": "#/definitions/ClusterKey"
        },
        "databaseType": {
          "description": "DatabaseType of the aerospike cluster (memory/ssd/performance)",
          "type": "string",
          "x-go-name": "DatabaseType",
          "example": "memory"
        },
        "name": {
          "description": "Name of aerospike cluster",
          "type": "string",
          "x-go-name": "Name",
          "example": "name"
        },
        "namespace": {
          "description": "Namespace of aerospike cluster",
          "type": "string",
          "x-go-name": "Namespace",
          "example": "default"
        },
        "options": {
          "$ref": "#/definitions/DatabaseOptions"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api/v1"
    },
...
    "AwsRegion": {
      "type": "object",
      "properties": {
        "continent": {
          "type": "string",
          "x-go-name": "Continent"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "region": {
          "type": "string",
          "x-go-name": "Region"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api-server/pkg/responses"
    },
    "AwsRegionsResponse": {
      "type": "object",
      "properties": {
        "regions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AwsRegion"
          },
          "x-go-name": "Regions"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api-server/pkg/responses"
    },
...
    "ClusterKey": {
      "description": "ClusterKey is the key to the cluster",
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api/v1"
    },
...
    "CreateAerospikeClusterRequest": {
      "description": "CreateAerospikeClusterRequest A request to create an aerospike cluster",
      "type": "object",
      "required": [
        "region",
        "name",
        "databaseType",
        "options"
      ],
      "properties": {
        "databaseType": {
          "description": "DatabaseType of the aerospike cluster (memory/ssd/performance)",
          "type": "string",
          "x-go-name": "DatabaseType",
          "example": "memory"
        },
        "name": {
          "description": "Name of aerospike cluster",
          "type": "string",
          "x-go-name": "Name",
          "example": "name"
        },
        "options": {
          "$ref": "#/definitions/DatabaseOptions"
        },
        "region": {
          "description": "KubernetesClusterName of workload cluster",
          "type": "string",
          "x-go-name": "Region",
          "example": "cluster-1"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api-server/pkg/requests"
    },
    "DatabaseOptions": {
      "description": "DatabaseOptions the mutable options of an aerospike cluster",
      "type": "object",
      "required": [
        "replicas"
      ],
      "properties": {
        "replicas": {
          "description": "Replicas of the aerospike cluster",
          "type": "integer",
          "format": "int32",
          "x-go-name": "Replicas",
          "example": 2
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api/v1"
    },
...
  },
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "BasicAuth": []
    }
  ]
}